requirements.txt
aiohttp
backoff
dependency-injector
pydantic
pyyaml